fix 64 bit literal parsing

+++ src/constant.cpp
@@ -54,7 +54,7 @@ std::pair<u64,Type*> compile_const_int_expression(Interloper& itl, AstNode* node
             const auto pool_slot = pool_slot_from_sym(sym);
             auto& section = pool_section_from_slot(itl.const_pool,pool_slot);
 
-            const u32 v = read_mem<u64>(itl.const_pool.buf,section.offset);
+            const u64 v = read_mem<u64>(itl.const_pool.buf,section.offset);
 
             return std::pair{v,type};
         }
@@ -66,7 +66,7 @@ std::pair<u64,Type*> compile_const_int_expression(Interloper& itl, AstNode* node
             const auto [value_left,type_left] = compile_const_int_expression(itl,bin_node->left);
             const auto [value_right,type_right] = compile_const_int_expression(itl,bin_node->right);
 
-            const u32 ans = value_left + value_right;
+            const u64 ans = value_left + value_right;
             Type* type = effective_arith_type(itl,type_left,type_right);
 
             return std::pair{ans,type};
@@ -79,7 +79,7 @@ std::pair<u64,Type*> compile_const_int_expression(Interloper& itl, AstNode* node
             const auto [value_left,type_left] = compile_const_int_expression(itl,bin_node->left);
             const auto [value_right,type_right] = compile_const_int_expression(itl,bin_node->right);
 
-            const u32 ans = value_left << value_right;
+            const u64 ans = value_left << value_right;
             Type* type = effective_arith_type(itl,type_left,type_right);
 
             return std::pair{ans,type};            
@@ -92,7 +92,7 @@ std::pair<u64,Type*> compile_const_int_expression(Interloper& itl, AstNode* node
             const auto [value_left,type_left] = compile_const_int_expression(itl,bin_node->left);
             const auto [value_right,type_right] = compile_const_int_expression(itl,bin_node->right);
 
-            const u32 ans = value_left >> value_right;
+            const u64 ans = value_left >> value_right;
             Type* type = effective_arith_type(itl,type_left,type_right);
 
             return std::pair{ans,type};            
@@ -105,7 +105,7 @@ std::pair<u64,Type*> compile_const_int_expression(Interloper& itl, AstNode* node
             const auto [value_left,type_left] = compile_const_int_expression(itl,bin_node->left);
             const auto [value_right,type_right] = compile_const_int_expression(itl,bin_node->right);
 
-            const u32 ans = value_left * value_right;
+            const u64 ans = value_left * value_right;
             Type* type = effective_arith_type(itl,type_left,type_right);
 
             return std::pair{ans,type};
@@ -118,7 +118,7 @@ std::pair<u64,Type*> compile_const_int_expression(Interloper& itl, AstNode* node
             const auto [value_left,type_left] = compile_const_int_expression(itl,bin_node->left);
             const auto [value_right,type_right] = compile_const_int_expression(itl,bin_node->right);
 
-            const u32 ans = value_left - value_right;
+            const u64 ans = value_left - value_right;
             Type* type = effective_arith_type(itl,type_left,type_right);
 
             return std::pair{ans,type};
@@ -137,7 +137,7 @@ std::pair<u64,Type*> compile_const_int_expression(Interloper& itl, AstNode* node
                 return std::pair{0,make_builtin(itl,builtin_type::void_t)}; 
             }
 
-            const u32 ans = value_left / value_right;
+            const u64 ans = value_left / value_right;
             Type* type = effective_arith_type(itl,type_left,type_right);
 
             return std::pair{ans,type};
